isolated the changes into the /test/e2e/accounts folder
