+++ test/e2e/accounts/utils.ts
@@ -1,0 +1,2 @@
+export const TEST_SNAPS_SIMPLE_KEYRING_WEBSITE_URL =
+  'https://metamask.github.io/snap-simple-keyring/0.2.3/';
