+++ test/e2e/snaps/enums.js
@@ -2,5 +2,5 @@ module.exports = {
   TEST_SNAPS_WEBSITE_URL:
     'https://metamask.github.io/snaps/test-snaps/0.38.0-flask.1/',
   TEST_SNAPS_SIMPLE_KEYRING_WEBSITE_URL:
-    'https://metamask.github.io/snap-simple-keyring/0.2.3/',
+    'https://metamask.github.io/snap-simple-keyring/0.2.2/',
 };
